refactor: extract `ScannedQrCodeDialog` from `ChannelFragment`
